fileShare/requirements.txt
asgiref==3.9.1
certifi==2025.7.9
cloudinary==1.44.1
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
mongoengine==0.29.1
PyJWT==2.9.0
pymongo==3.11.4
python-dotenv==1.1.1
python-magic==0.4.27
pytz==2025.2
redis==6.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
